requirements.txt
# ML
torch
torchvision

# YOLO Model
ultralytics

# Metrics
torchmetrics
pycocotools

# Utils
requests
pyyaml
pandas
numpy
pillow

